+++ example/lib/pages/complex/pull_to_refresh_chat.dart
@@ -3,6 +3,7 @@ import 'package:extended_list/extended_list.dart';
 import 'package:ff_annotation_route_library/ff_annotation_route_library.dart';
 import 'package:flutter/material.dart';
 import 'package:pull_to_refresh_notification/pull_to_refresh_notification.dart';
+
 @FFRoute(
   name: 'fluttercandies://PullToRefreshChat',
   routeName: 'PullToRefreshChat',